Added method to create Pantry Document in MongoDB TODO: need verify ownerID

+++ Progetto-Gruppo-1/Pantry/src/main/java/it/unisannio/ingsw24/pantry/persistence/PantryDAO.java
@@ -25,7 +25,7 @@ public interface PantryDAO {
 
     boolean closeConnection();
 
-    Long createPantry(Pantry p);
+    int createPantry(Pantry p);
 
     List<Pantry> getPantries(int id);
